--- web/html/src/manager/organizations/index.js
@@ -1,3 +1,0 @@
-export default {
-  'organizations/config-channels/org-config-channels': () => import('./config-channels/org-config-channels')
-}
