+++ web/html/src/manager/organizations/index.ts
@@ -1,0 +1,3 @@
+export default {
+  "organizations/config-channels/org-config-channels": () => import("./config-channels/org-config-channels"),
+};
